Working on scanner page
Making first steps towards creating a functioning scanner screen, I have all the UI, added some buttons and functionality to various features. When typing in an id, there is a default image when the input box is empty, an error image when there is no id found, and if there is an image for that id, it displays it successfully.