Feedback, formatting, users section working
Finishing the last of the feedback from the update stock assignment. Removing an image that was unused, removing 70+ lines of comments, formatting some if statements and adding spacing around methods and code blocks, simplifying some validation methods, added a getUser method similar to the getItem in sqlutils, added a variable to the user class for a student picture, removed unneeded files, similar to the image on the scanner page, there is a responsive image for the users page.