Fixing images, images should function properly
Updating several methods like addItem and updateItem, to include image paths so everything should have functioning images. Making createImage shorter to work better, adding a file chooser that allows the user to choose an image, and set the image for the item that is currently selected to the chosen image, it works and updates after every key typed on several pages.