Lots of ui and naming changes
Created an adjustedItem class, a functioning table that responds to what user is clicked. After a user makes a change to the table, you can see what transactions they made and the adjustment for each item in the transaction table. I made the default account for the transactions the “sean riley”. Changed some of the region names to make more sense. I'm in the process of removing the email field from the login page. I removed all instances of it and disabled it but I haven't posted anything yet. Changed some variable names in SQLUtils file, created an addTransaction and getTransaction methods.