Code simplification, comments, spacing, feedback, altering and finishing user transactions requirements
Finishing the user transactions and item adjustments requirements for the project. Changing the system to work better and how Mr Riley wanted, every item had a transaction number, now there is one transaction per finalize button press. Code simplification and formatting, lots of comments and spacing, changing some sql slightly. Moved regexValidation method to utils because I'm using it on the login page and on the user create screen in the dashboard page. Removing the userExists method and making things that use it work with getUser instead. Making the users section of the dashboard page inaccessible unless the user has a grade level of -1, meaning they are an admin.